code complete but possibly buggy
